third_party/blink/web_tests/platform/generic/fast/block/align-bigger-child-expected.txt

PASS rtl_in_ltr_left.getBoundingClientRect().left is 0
PASS ltr_in_ltr_left.getBoundingClientRect().left is 0
PASS rtl_in_ltr_right.getBoundingClientRect().left is 0
PASS ltr_in_ltr_right.getBoundingClientRect().left is 0
PASS rtl_in_ltr_center.getBoundingClientRect().left is 0
PASS ltr_in_ltr_center.getBoundingClientRect().left is 0
PASS rtl_in_rtl_right.getBoundingClientRect().right is 100
PASS ltr_in_rtl_right.getBoundingClientRect().right is 100
PASS rtl_in_rtl_left.getBoundingClientRect().right is 100
PASS ltr_in_rtl_left.getBoundingClientRect().right is 100
PASS rtl_in_rtl_center.getBoundingClientRect().right is 100
PASS ltr_in_rtl_center.getBoundingClientRect().right is 100
PASS successfullyParsed is true

TEST COMPLETE
